clip-service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
pydantic==2.9.0
pillow==10.4.0
httpx==0.27.0
numpy==1.26.4
torch==2.4.0
transformers==4.44.0
